Once-over tests and cleanup
