+++ template_test.go
@@ -3,8 +3,6 @@ package main
 import (
 	"bytes"
 	"os"
-	"reflect"
-	"sort"
 	"strings"
 	"testing"
 
@@ -541,46 +539,6 @@ func TestHashCode_returnsValue(t *testing.T) {
 	expected := "Template|/foo/bar/blitz.ctmpl"
 	if template.HashCode() != expected {
 		t.Errorf("expected %q to equal %q", template.HashCode(), expected)
-	}
-}
-
-func TestServiceList_sorts(t *testing.T) {
-	a := util.ServiceList{
-		&util.Service{Node: "frontend01", ID: "1"},
-		&util.Service{Node: "frontend01", ID: "2"},
-		&util.Service{Node: "frontend02", ID: "1"},
-	}
-	b := util.ServiceList{
-		&util.Service{Node: "frontend02", ID: "1"},
-		&util.Service{Node: "frontend01", ID: "2"},
-		&util.Service{Node: "frontend01", ID: "1"},
-	}
-	c := util.ServiceList{
-		&util.Service{Node: "frontend01", ID: "2"},
-		&util.Service{Node: "frontend01", ID: "1"},
-		&util.Service{Node: "frontend02", ID: "1"},
-	}
-
-	sort.Stable(a)
-	sort.Stable(b)
-	sort.Stable(c)
-
-	expected := util.ServiceList{
-		&util.Service{Node: "frontend01", ID: "1"},
-		&util.Service{Node: "frontend01", ID: "2"},
-		&util.Service{Node: "frontend02", ID: "1"},
-	}
-
-	if !reflect.DeepEqual(a, expected) {
-		t.Fatal("invalid sort")
-	}
-
-	if !reflect.DeepEqual(b, expected) {
-		t.Fatal("invalid sort")
-	}
-
-	if !reflect.DeepEqual(c, expected) {
-		t.Fatal("invalid sort")
 	}
 }
